get rid of \ characters in the texts and reorganize java packages
