+++ app/src/main/java/com/example/android/famousWomen/Glide/ImageDownload.java
@@ -1,9 +1,11 @@
-package com.example.android.famousWomen;
+package com.example.android.famousWomen.Glide;
 
 import android.content.Context;
 import android.widget.ImageView;
 
 import com.bumptech.glide.Glide;
+import com.example.android.famousWomen.Glide.GlideApp;
+import com.example.android.famousWomen.R;
 //import com.squareup.picasso.Picasso;
 
 /**
