+++ app/src/main/java/com/example/android/famousWomen/Modal/Woman.java
@@ -13,10 +13,8 @@
  * See the License for the specific language governing permissions and
  * limitations under the License.
  */
-package com.example.android.famousWomen;
+package com.example.android.famousWomen.Modal;
 
-import android.content.Context;
-import android.content.res.Resources;
 import android.os.Parcel;
 import android.os.Parcelable;
