+++ app/src/main/java/com/example/android/famousWomen/RecyclerView/ItemClickListener.java
@@ -1,4 +1,4 @@
-package com.example.android.famousWomen;
+package com.example.android.famousWomen.RecyclerView;
 
 import android.view.View;
