+++ app/src/main/java/com/example/android/famousWomen/RecyclerView/WomenHolder.java
@@ -1,9 +1,11 @@
-package com.example.android.famousWomen;
+package com.example.android.famousWomen.RecyclerView;
 
 import android.support.v7.widget.RecyclerView;
 import android.view.View;
 import android.widget.ImageView;
 import android.widget.TextView;
+
+import com.example.android.famousWomen.R;
 
 /**
  * Created by narcy on 25.02.2018.
@@ -21,10 +23,10 @@ public class WomenHolder extends RecyclerView.ViewHolder implements View.OnClick
     public WomenHolder(View itemView) {
         super(itemView);
 
-        this.nameTextView = (TextView) itemView.findViewById(R.id.name_text_view);
-        this.professionTextView = (TextView) itemView.findViewById(R.id.profession_text_view);
-        this.imageView = (ImageView) itemView.findViewById(R.id.image);
-        this.flagImage = (ImageView) itemView.findViewById(R.id.flag);
+        this.nameTextView = itemView.findViewById(R.id.name_text_view);
+        this.professionTextView = itemView.findViewById(R.id.profession_text_view);
+        this.imageView = itemView.findViewById(R.id.image);
+        this.flagImage = itemView.findViewById(R.id.flag);
 
 
         itemView.setOnClickListener(this);
