+++ app/src/main/res/layout-land/details_view.xml
@@ -2,7 +2,7 @@
 <android.support.design.widget.CoordinatorLayout xmlns:android="http://schemas.android.com/apk/res/android"
     xmlns:app="http://schemas.android.com/apk/res-auto"
     xmlns:tools="http://schemas.android.com/tools"
-    tools:context=".DetailsActivity"
+    tools:context=".Activities.DetailsActivity"
     android:id="@+id/main_content"
     android:layout_width="match_parent"
     android:layout_height="match_parent"
