+++ app/src/main/res/layout/about_application.xml
@@ -6,7 +6,7 @@
     android:layout_width="match_parent"
     android:layout_height="match_parent"
     android:fitsSystemWindows="true"
-    tools:context=".QuizActivity">
+    tools:context=".Activities.QuizActivity">
 
     <android.support.design.widget.AppBarLayout
         android:id="@+id/appbar"
